+++ app/src/main/java/de/dreier/mytargets/fragments/EnvironmentFragment.java
@@ -121,8 +121,8 @@ public class EnvironmentFragment extends FragmentBase {
     private void onSave() {
         Environment e = new Environment();
         e.weather = weather;
-        e.windSpeed = (int) binding.windSpeed.getSelectedItem().getId();
-        e.windDirection = (int) binding.windDirection.getSelectedItem().getId();
+        e.windSpeed = (int) (long) binding.windSpeed.getSelectedItem().getId();
+        e.windDirection = (int) (long) binding.windDirection.getSelectedItem().getId();
         e.location = binding.location.getText().toString();
         listener.onItemSelected(Parcels.wrap(e));
     }
